requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async
httpx>=0.26.0
aiosqlite>=0.19.0

# YouTube/Download
yt-dlp>=2024.1.0

# Transcription
openai-whisper>=20231117

# Vector Database & Embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# LLM
groq>=0.4.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Type checking
mypy>=1.8.0

# Linting
ruff>=0.1.0
